backend/services/pronunciation-feedback/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.9
azure-cognitiveservices-speech>=1.34.0
pydantic-settings>=2.0.0
